Function moras::modules::riscv::middleware::backend_api::syscall_output_print

pub fn syscall_output_print(pathname: &str, output: &str) -> Result<(), String>

Emits a print syscall output event to the frontend.

pathname: Identifier for the tab to which the output should be sent.

output: Output to be printed.

Returns Result indicating the success or failure of the event emission.

This function will emit a front_syscall_print event to the frontend, and the payload is a SyscallOutput containing the filepath and output to be printed.

SyscallOutput:

filepath: string

data: string