Struct moras::remote::server::RpcServerImpl

pub struct RpcServerImpl {
    port: AtomicU16,
    tokio_runtime: Runtime,
    server_handle: Option<JoinHandle<()>>,
    shared_handler: Arc<Mutex<ServerHandle>>,
}

Fields
port: AtomicU16tokio_runtime: Runtimeserver_handle: Option<JoinHandle<()>>shared_handler: Arc<Mutex<ServerHandle>>
Implementations

impl RpcServerImpl

pub fn start_service( &mut self, cur_tab_name: String, tab_map: Ptr<TabMap> ) -> Result<(), Box<dyn Error>>

Start the service with a tab map.

If the server is already running, return an error.

pub fn stop_service(&mut self)

pub fn set_port(&mut self, port: u16) -> Result<(), Box<dyn Error>>

Set the port for the server.

If the server is already running, return an error.

pub fn change_password(&mut self, new_password: &str)

Change the password for the server.

pub fn is_running(&self) -> bool

Check if the server is running.

pub fn get_port(&self) -> u16

Get the port of the server.

pub fn set_host_cursor(&mut self, row: u64, col: u64)

Update the host cursor position.

row: The row of the cursor.

col: The column of the cursor. TODO: add out of range check