Module moras::modules::riscv::middleware::frontend_api

This module provides API functions for the frontend. Could be used by invoke in EMCAScript

Re-exports

pub use __cmd__create_tab;

pub use __cmd__close_tab;

pub use __cmd__change_current_tab;

pub use __cmd__set_cursor;

pub use __cmd__modify_current_tab;

pub use __cmd__read_tab;

pub use __cmd__write_tab;

pub use __cmd__set_return_data_range;

pub use __cmd__assembly;

pub use __cmd__dump;

pub use __cmd__run;

pub use __cmd__debug;

pub use __cmd__step;

pub use __cmd__reset;

pub use __cmd__undo;

pub use __cmd__set_breakpoint;

pub use __cmd__remove_breakpoint;

pub use __cmd__syscall_input;

pub use __cmd__update_assembler_settings;

pub use __cmd__start_share_server;

pub use __cmd__authorize;

pub use __cmd__stop_share_server;

Functions

assembly

Assembles the code in the currently active tab.

authorize

Authorize and connect to a remote RPC server as client.

change_current_tab

Changes the current tab to the one specified by the new path.

close_tab

Closes the tab associated with the given file path.

create_tab

Creates a new tab with content loaded from a specified file path.

debug

Run the code in the currently active tab in debug mode.

dump

Placeholder for a function to dump data from all tabs.

modify_current_tab

Updates the content of the tab associated with the given file path.

parse 🔒

helper function

read_tab

Reads the content of a tab from the file at the specified path.

remove_breakpoint

Removes a breakpoint at a specified line in the code of the current tab.

reset

Resets the state of the currently active tab’s simulator.

run

Run the code in the currently active tab in normal mode(won’t stop at exist break point).

set_breakpoint

Sets a breakpoint at a specified line in the code of the current tab.

set_cursor

Sets the cursor position in the tab associated with the given file path. This could be useful for live sharing of code.

set_return_data_range

Sets the range [start, end] of data to be returned by the assembly and simulator for the current tab. Default is [0, 0].

start_share_server

Starts the RPC server for the current tab.

step

Steps through the code in the currently active tab.

stop_share_server

Stop the share server for the current tab.

syscall_input

Send a syscall input to current tab’s simulator.

undo

Undoes the last instruction for current activate tab’s simulator.

update_assembler_settings

Updates the assembler settings for the current tab.

write_tab

Writes data in specific tab to the file file path.